+++ client/jetpack-cloud/sections/overview/controller.tsx
@@ -1,7 +1,7 @@
 import { type Callback, type Context } from '@automattic/calypso-router';
 import ContentSidebar from 'calypso/jetpack-cloud/components/content-sidebar';
 import Overview from 'calypso/jetpack-cloud/sections/overview/primary/overview';
-import OverviewSidebar from 'calypso/jetpack-cloud/sections/overview/sidebar';
+import OverviewSidebar from 'calypso/jetpack-cloud/sections/overview/primary/sidebar';
 import JetpackManageSidebar from 'calypso/jetpack-cloud/sections/sidebar-navigation/jetpack-manage';
 
 const setSidebar = ( context: Context ): void => {
